Add: status message

+++ NoMoreVanguard.bat
@@ -33,34 +33,44 @@ for %%a in ("installer.exe", "log-uploader.exe", "vgc.exe", "vgc.ico", "vgk.sys"
 
 :Toggle
 if exist "vgk.sys" (
-    REM Stops Vanguard services, renames key files, and deletes Vanguard logs
-    echo Disabling Vanguard...
-    sc config vgc start= disabled >nul 2>&1
-    sc config vgk start= disabled >nul 2>&1
-    net stop vgc >nul 2>&1
-    net stop vgk >nul 2>&1
-    taskkill /f /im vgtray.exe >nul 2>&1
-    for %%a in (*) do (
-        ren "%%a" "%%a.bak"
-    )
-    del /q "Logs"
-) else (
-    REM Reverts changes made by disable function and reinstates services, then restarts the system after 30 seconds w/ countdown
-    echo Enabling Vanguard...
-    for %%a in (*) do (
-        ren "%%a" "%%~na"
-    )
-    sc config vgc start= demand
-    sc config vgk start= system
-    for /l %%b in (30 -1 1) do (
-        cls
-        echo Restarting in: %%b Seconds - [C]ancel / [R]estart Now
-        for /f "delims=" %%c in ('Choice /T 1 /N /C:CSRW /D W') do (
-            if %%c==C goto :Toggle
-            if %%c==R shutdown /r /f /t 00
+    echo Vanguard is currently enabled. Would you like to disable it? [Y/N]
+    choice /c YN /n
+    if errorlevel 2 goto :eof
+    if errorlevel 1 (
+        REM Stops Vanguard services, renames key files, and deletes Vanguard logs
+        echo Disabling Vanguard...
+        sc config vgc start= disabled >nul 2>&1
+        sc config vgk start= disabled >nul 2>&1
+        net stop vgc >nul 2>&1
+        net stop vgk >nul 2>&1
+        taskkill /f /im vgtray.exe >nul 2>&1
+        for %%a in (*) do (
+            ren "%%a" "%%a.bak"
         )
+        del /q "Logs"
     )
-    shutdown /r /f /t 00
+) else (
+    echo Vanguard is currently disabled. Would you like to enable it? [Y/N]
+    choice /c YN /n
+    if errorlevel 2 goto :eof
+    if errorlevel 1 (
+        REM Reverts changes made by disable function and reinstates services, then restarts the system after 30 seconds w/ countdown
+        echo Enabling Vanguard...
+        for %%a in (*) do (
+            ren "%%a" "%%~na"
+        )
+        sc config vgc start= demand
+        sc config vgk start= system
+        for /l %%b in (30 -1 1) do (
+            cls
+            echo Restarting in: %%b Seconds - [C]ancel / [R]estart Now
+            for /f "delims=" %%c in ('Choice /T 1 /N /C:CSRW /D W') do (
+                if %%c==C goto :Toggle
+                if %%c==R shutdown /r /f /t 00
+            )
+        )
+        shutdown /r /f /t 00
+    )
 )
 
 exit
